requirements.txt
asgiref==3.3.1
Django==3.1.3
django-crispy-forms==1.11.1
djangorestframework==3.12.2
mysqlclient
Pillow==8.0.1
sqlparse==0.4.1
appdirs==1.4.4
distlib==0.3.1
filelock==3.0.12
